fixed broker file structure and docs
